+++ packages/react-tinacms-github/src/github-client/github-client.ts
@@ -27,12 +27,15 @@ export interface GithubClientOptions {
   authCallbackRoute: string
   baseRepoFullName: string
   baseBranch?: string
+  authScope?: AuthScope
 }
 
 export interface Branch {
   name: string
   protected: boolean
 }
+
+export type AuthScope = 'public_repo' | 'repo'
 
 export class GithubClient {
   static WORKING_REPO_COOKIE_KEY = 'working_repo_full_name'
@@ -43,6 +46,7 @@ export class GithubClient {
   baseBranch: string
   clientId: string
   authCallbackRoute: string
+  authScope: AuthScope
 
   constructor({
     proxy,
@@ -50,17 +54,19 @@ export class GithubClient {
     authCallbackRoute,
     baseRepoFullName,
     baseBranch = 'master',
+    authScope = 'public_repo',
   }: GithubClientOptions) {
     this.proxy = proxy
     this.baseRepoFullName = baseRepoFullName
     this.baseBranch = baseBranch
     this.clientId = clientId
     this.authCallbackRoute = authCallbackRoute
+    this.authScope = authScope
     this.validate()
   }
 
   authenticate() {
-    return authenticate(this.clientId, this.authCallbackRoute)
+    return authenticate(this.clientId, this.authCallbackRoute, this.authScope)
   }
 
   isAuthenticated() {
@@ -133,9 +139,11 @@ export class GithubClient {
       return forkName
     }
 
-    this.setCookie(GithubClient.WORKING_REPO_COOKIE_KEY, this.baseRepoFullName)
-
     return this.baseRepoFullName
+  }
+
+  setWorkingRepoFullName(repoFullName: string) {
+    this.setCookie(GithubClient.WORKING_REPO_COOKIE_KEY, repoFullName)
   }
 
   get branchName(): string {
@@ -145,9 +153,7 @@ export class GithubClient {
       return branchName
     }
 
-    this.setCookie(GithubClient.HEAD_BRANCH_COOKIE_KEY, this.baseBranch)
-
-    return this.branchName
+    return this.baseBranch
   }
 
   setWorkingBranch(branch: string) {