+++ packages/@tinacms/react-screens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-screens@0.1.2...@tinacms/react-screens@0.1.3-alpha.0) (2020-05-08)
+
+**Note:** Version bump only for package @tinacms/react-screens
+
+
+
+
+
 ## [0.1.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-screens@0.1.2-alpha.0...@tinacms/react-screens@0.1.2) (2020-05-04)
 
 **Note:** Version bump only for package @tinacms/react-screens