+++ packages/react-tinacms-inline/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.7-alpha.3](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.7-alpha.2...react-tinacms-inline@0.4.7-alpha.3) (2020-05-08)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
+## [0.4.7-alpha.2](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.7-alpha.1...react-tinacms-inline@0.4.7-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
+## [0.4.7-alpha.1](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.7-alpha.0...react-tinacms-inline@0.4.7-alpha.1) (2020-05-08)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
+## [0.4.7-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.6...react-tinacms-inline@0.4.7-alpha.0) (2020-05-06)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
 ## [0.4.6](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.6-alpha.0...react-tinacms-inline@0.4.6) (2020-05-04)
 
 **Note:** Version bump only for package react-tinacms-inline